refactor: Consistent usage of SerializableEdwardsPoint in crypto_shared/types.rs
